spline #306 + builder.read/write package
