--- spark/agent/.gitignore
@@ -1,1 +1,0 @@
-src/test/resources/data
