+++ spark/agent/src/main/scala/za/co/absa/spline/harvester/builder/read/ReadNodeBuilder.scala
@@ -14,32 +14,28 @@
  * limitations under the License.
  */
 
-package za.co.absa.spline.harvester.builder
+package za.co.absa.spline.harvester.builder.read
 
-import org.apache.spark.sql.SaveMode
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import za.co.absa.spline.harvester.ComponentCreatorFactory
 import za.co.absa.spline.harvester.ModelConstants.OperationParams
-import za.co.absa.spline.producer.rest.model.WriteOperation
+import za.co.absa.spline.harvester.builder.OperationNodeBuilder
+import za.co.absa.spline.producer.rest.model.ReadOperation
 
-class WriteNodeBuilder(command: WriteCommand)
+class ReadNodeBuilder
+(val command: ReadCommand)
   (implicit val componentCreatorFactory: ComponentCreatorFactory)
   extends OperationNodeBuilder {
 
-  override protected type R = WriteOperation
-  override val operation: LogicalPlan = command.query
+  override protected type R = ReadOperation
+  override val operation: LogicalPlan = command.operation
 
-  override def build(): WriteOperation = {
-    val Seq(uri) = command.sourceIdentifier.uris
-    WriteOperation(
-      outputSource = uri,
-      append = command.mode == SaveMode.Append,
-      id = id,
-      childIds = childIds,
-      schema = Some(outputSchema),
-      params = command.params ++ Map(
-        OperationParams.Name -> command.nodeName,
-        OperationParams.DestinationType -> command.sourceIdentifier.format)
-    )
-  }
+  override def build(): ReadOperation = ReadOperation(
+    inputSources = command.sourceIdentifier.uris,
+    id = id,
+    schema = Some(outputSchema),
+    params = Map(
+      OperationParams.Name -> operation.nodeName,
+      OperationParams.SourceType -> command.sourceIdentifier.format
+    ))
 }
